Move govuk scripts into actual scripts
Having them as shell functions is inconvenient because the shell
config has to be reloaded every time a change is made.  The functions
have grown too big for that to be a nice workflow.

Also fix the ES replication, which wasn't quite working.  It now
depends on some outstanding changes to govuk-docker.